provide target 'schema-upgrade'
Once in a while a new ::Schema contains real
SQL table changes. The ::Schema transports
the respective upgrade scripts to migrate the
DB from old version to new version.

make schema-upgrade

now calls the neccessary cmdline to do that
upgrade.

It is not done implicitely to enforce humans to
think before that, e.g. about backups and or
neccessity.